Move RPC types in to RpcModels folder
