+++ WalletWasabi/BitcoinCore/RpcClientBase.cs
@@ -2,6 +2,7 @@ using NBitcoin;
 using NBitcoin.RPC;
 using System;
 using System.Threading.Tasks;
+using WalletWasabi.BitcoinCore.RpcModels;
 using WalletWasabi.Helpers;
 
 namespace WalletWasabi.BitcoinCore
