+++ WalletWasabi/BitcoinCore/RpcModels/VerboseTransactionInfo.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
+using NBitcoin;
 
-namespace NBitcoin.RPC
+namespace WalletWasabi.BitcoinCore.RpcModels
 {
 	public class VerboseTransactionInfo
 	{
